Add chat auto scroll feature using Static() widget
This feature uses a "hidden" `Static(disabled=True)` widget added to the message container to control automatic scrolling of new messages based on the visibility of the widget (`can_view()`). The resulting behavior is that automatic scrolling is enabled for all new chats. To disable automatic scrolling, simply scroll up so that the most resent message is out of focus. To re-enable auto scrolling, scroll back down.